% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add classification data to data frame with case data}
\usage{
add(what, to, from, id, date, days = NULL, ind = NULL, tech_names = FALSE,
  sort = TRUE)
}
\arguments{
\item{what}{to add? classification scheme of type classcodes to classify by}

\item{to}{where should it be added? A data frame with at least two columns,
one with case (patient) identification (column name specified by argument id)
and one with a date of interest (column name specified by argument date)}

\item{from}{where do we find the class codes?
Object with code data for which \code{\link{is.codedata}} is \code{TRUE}}

\item{id, date, days}{arguments passed to \code{\link{codify}}}

\item{ind}{control possible inclusion of index vector. Set to \code{FALSE} if
no index should be calculated, otherwise a value passed to argument \code{by}
in function \code{\link{index}}. It is here also possible to include several
indices as a character vector.}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{by}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.}

\item{sort}{logical. Should the result be sorted by the 'id' column?}
}
\value{
data frame with all data from input data frame \code{to} combined
with added class (and possibly index) data.
}
\description{
This function can be used as a convenient short cut for the otherwise common
multi step procedure based on \code{\link{codify}}, \code{\link{classify}},
\code{\link{index}} and then combining the input data with the new result.
}
\examples{
add("elix_icd10", to = ex_people, from = ex_icd10,
  id = "name", date = "surgery")
}
