% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classcodes.R
\name{all_classcodes}
\alias{all_classcodes}
\title{Summary data for all default classcodes object in the package}
\usage{
all_classcodes()
}
\value{
\link[tibble:tibble]{tibble::tibble} with columns describing all default classcodes
objects from the package.
}
\description{
Tabulate object names and list all related versions of implemented regular
expressions
and index weights.
}
\examples{
all_classcodes()
}
\seealso{
Other classcodes: 
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebook}()},
\code{\link{print.classcodes}()},
\code{\link{print.classified}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
