% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet.list}
\alias{alphabet.list}
\title{List available alphabets}
\usage{
alphabet.list()
}
\value{
A vector of names
}
\description{
List available alphabets
}
\examples{
alphabet.list()
}
\keyword{alphabet}
\keyword{nato,}

