% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet}
\alias{alphabet}
\title{Return a phonetic alphabet}
\usage{
alphabet(alphabet = "nato")
}
\arguments{
\item{alpabet}{The alpabet, for example "nato"}
}
\value{
A vector
}
\description{
Return a phonetic alphabet
}
\examples{
nato <- alphabet("nato")
}
\keyword{alphabet}
\keyword{nato,}

