% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{as.codedata}
\alias{as.codedata}
\alias{as.codedata.data.table}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ...)

\method{as.codedata}{data.table}(
  x,
  id = "id",
  code = "code",
  code_date = if (utils::hasName(x, "code_date")) "code_date" else NULL,
  period = NULL,
  .setkeys = TRUE,
  alnum = FALSE,
  ...
)

is.codedata(x)
}
\arguments{
\item{x}{data table/frame with columns \code{id} (character), \code{code},
and optional \code{code_date} (Date).}

\item{...}{arguments passed between methods}

\item{id, code}{column names (from \code{x}) with case id:s (character) and
codes of interest.}

\item{code_date}{optional column name (from \code{x}) with relevant dates
corresponding to each code. If there is a column named \code{code_date} this will
be used by default.}

\item{period}{optional date vector of length two with lower and upper limits
for inclusion period}

\item{.setkeys}{should output be indexed (Beneficial for large data sets.)}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters?}
}
\value{
\link{data.table} with columns: \code{id} (character), \code{code} and
optionally \code{code_date} (Date)
Possible additional columns from \code{x} are preserved.
}
\description{
Make structured code data set.
}
\examples{

x <-
  data.frame(
    id               = "1",
    code_date        = as.Date("2017-01-01"),
    code             = "a",
    stringsAsFactors = FALSE
  )
as.codedata(x)

# Ignore codes for dates outside specified limits
y <-
  data.frame(
    id               = "2",
    code_date        = as.Date("3017-02-02"),
    code             = "b",
    stringsAsFactors = FALSE
  )
z <- rbind(x, y)
as.codedata(z, period = c(as.Date("1970-01-01"), Sys.Date()))

}
\concept{codedata}
