% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.keyvalue.R
\name{as.keyvalue.classcodes}
\alias{as.keyvalue.classcodes}
\title{Make keyvalue object from classcodes object}
\usage{
\method{as.keyvalue}{classcodes}(x, coding, ...)
}
\arguments{
\item{x}{classcodes object}

\item{coding}{either a vector with codes from the original classification,
or a name (character vector of length one) keyvalue object from package
"decoder" (for example "icd10cm" or "atc")}

\item{...}{arguments passed to \code{\link[=summary.classcodes]{summary.classcodes()}}}
}
\description{
S3-method for generic \code{\link[decoder:keyvalue]{decoder::as.keyvalue()}}
}
\examples{
# List all codes with corresponding classes as recognized by the Elixhauser
# comorbidity classification according to the Swedish version of the
# international classification of diseases version 10 (ICD-10-SE)
decoder::as.keyvalue(elixhauser, "icd10se")
}
\seealso{
Other helper: 
\code{\link{dates_within}()},
\code{\link{filter_dates}()},
\code{\link{spss2date}()}
}
\concept{helper}
