% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_text}
\alias{clean_text}
\title{Make clean text with only lowercase alphanumeric characters and "_"}
\usage{
clean_text(x_name, x)
}
\arguments{
\item{x_name}{Name of object to use as prefix}

\item{x}{character vector}
}
\value{
character vector of the same length as \code{x}
}
\description{
Make clean text with only lowercase alphanumeric characters and "_"
}
\keyword{internal}
