% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{print.codebook}
\alias{codebooks}
\title{codebook(s) for classcodes object}
\usage{
codebook(object, coding, ..., file = NULL)

\method{print}{codebook}(x, ...)

codebooks(..., file = NULL)
}
\arguments{
\item{object}{classcodes object}

\item{coding}{either a vector with codes from the original classification,
or a name (character vector of length one) of a keyvalue object
from package "decoder" (for example "icd10cm" or "atc")}

\item{...}{Additional arguments for each function:
\itemize{
\item \code{codebook()}: arguments passed to \code{\link[=summary.classcodes]{summary.classcodes()}}
\item \code{codebooks()}: multiple named outputs from \code{\link[=codebook]{codebook()}}
\item \code{print.codebook()}: arguments passed to \code{tibble:::print.tbl()}
}}

\item{file}{name/path to Excel file for data export}

\item{x}{output from \code{codebook()}}
}
\value{
Functions are primarily called for their side effects (exporting data to
Excel or printing to screen). In addition:
\itemize{
\item \code{codebook()}returns list of data frames describing relationship
between groups and individual codes
\item \code{codebooks()} returns a concatenated list with output from \code{codebook()}.
Only one 'README' object is kept however and renamed as such.
\item \code{print.codebook()}returns \code{x} (invisible)
}
}
\description{
\code{\link[=summary.classcodes]{summary.classcodes()}} and \code{\link[=visualize.classcodes]{visualize.classcodes()}} are used to
summarize/visualize classcodes in R. A codebook, on the other hand,
is an exported summary
saved in an Excel spreadsheet to use in collaboration with non R-users.
Several codebooks might be combined into a single Excel document with
several sheets (one for each codebook).
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{print.classcodes}()},
\code{\link{print.classified}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
