% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebooks}
\alias{codebooks}
\title{Combine codebooks for classcodes objects}
\usage{
codebooks(..., file = NULL)
}
\arguments{
\item{...}{named output from \code{\link{codebook}}}

\item{file}{name of Excel file for data export}
}
\value{
Concatenated list with output from \code{\link{codebook}}.
Only one 'README' object is kept however and renamed as such.
}
\description{
Combine codebooks for classcodes objects
}
\examples{
c1 <- codebook(elixhauser, "icd10")
c2 <- codebook(elixhauser, "icd9cmd",
  cc_args = list(regex = "regex_icd9cm_enhanced")
  )
codebooks(elix_icd10 = c1, elix_icd9cm = c2)
}
\seealso{
codebook
}
