% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cols}
\alias{cols}
\title{Return all columns from x with names matching "find"}
\usage{
cols(find, x)
}
\arguments{
\item{find}{character vector with names to match}

\item{x}{matrix}
}
\description{
Return all columns from x with names matching "find"
}
\keyword{internal}
