% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{cps_icd10}
\alias{cps_icd10}
\title{Classcodes for the comorbidity-polypharmacy score (CPS) based on ICD-10 codes}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{group}{comorbidity groups, either "ordinary" for most ICD-10-codes or
  "special" for codes beginning with "UA", "UB" and "UP"}
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
  \item{only_ordinary}{index weights, 1 for ordinary and 0 for special}
}}
\source{
Stawicki, Stanislaw P., et al.
"Comorbidity polypharmacy score and its clinical utility: A pragmatic
practitioner's perspective." Journal of emergencies, trauma, and shock 8.4
(2015): 224.
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4626940/}
}
\usage{
cps_icd10
}
\description{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\seealso{
Other default classcodes: \code{\link{ae}},
  \code{\link{charlson_icd10}}, \code{\link{elix_icd10}},
  \code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
