% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill}
\alias{distill}
\title{Title}
\usage{
distill(x, y, id = "id", date, years = 1, keep_names = FALSE)
}
\arguments{
\item{x}{data.frame with at least two columns, one with case (patient)
identification (column name specified by argument \code{id}) and one with a
date of interest (column name specified by argument \code{date})}

\item{y}{object of class \code{\link{codedata}}}

\item{id}{name of column in \code{x} containing case (patient) identification}

\item{date}{name of column in \code{x} date of interest}

\item{years}{number of years before \code{date} during which codes from
\code{y} are valid}
}
\value{
Object of class \code{tbl_df} with rows from \code{x} compatible
with \code{y} and an additional attribute "id" specifynig the name of the id
column (as specified by argument \code{id})
}
\description{
Title
}
\examples{
distill(ex_people, ex_icd10, id = "name", date = "surgery")
}

