% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{elix_icd10}
\alias{elix_icd10}
\title{Classcodes for Elixhauser based on ICD-10 codes}
\format{A data frame with 31 rows and 2 variables:
\describe{
  \item{group}{comorbidity groups}
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
}}
\source{
Quan Hude et. al. (2005). Coding algorithms for defining comorbidities
in ICD-9-CM and ICD-10 administrative data. Medical care, 1130-1139.
\url{www.jstor.org/stable/3768193}
}
\usage{
elix_icd10
}
\description{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\keyword{datasets}

