% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{ex_atc}
\alias{ex_atc}
\title{Example data for random ATC codes}
\format{
Data frames with 100 rows and 2 variables:
\describe{
\item{name}{random person names}
\item{atc}{Random codes from the Anatomic Therapeutic Chemical
classification (ATC) system.}
\item{prescription}{random dates of prescription of medications with c
oresponding ATC codes}
}
}
\usage{
ex_atc
}
\description{
Example data for fictive people to use for testing and in examples.
}
\seealso{
Other example data: 
\code{\link{ex_icd10}},
\code{\link{ex_people}}
}
\concept{example data}
\keyword{datasets}
