% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{ex_pardata}
\alias{ex_pardata}
\title{Example NPR data with random codes assigned to random people}
\format{Data frames with 1,000 rows and 18 columns:
\describe{
  \item{lpnr}{Random names of people}
  \item{date}{random dates for inpatient registration}
  \item{hdia}{main diagnose given by ICD-10}
  \item{bdia(1, ..., 14)}{additional diagnoses for each visit}
}}
\usage{
ex_pardata
}
\description{
Eictive ICD-10-codes in the format as delivered by the
National Patient Register (NPR).
}
\details{
Note that a real NPR data set would have "bdia" codes filled in order.
The situation with no "bdia1" but with a "bdia2" is therefore not
possible in a real data set.
}
\seealso{
\code{\link{ex_people}}, \code{\link{ex_icd10}}
}
\keyword{datasets}
