% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{hip_adverse_events_icd10_old}
\alias{hip_adverse_events_icd10_old}
\title{Old classcodes for adverse events after hip arthroplasty based on ICD-10 codes}
\format{A data frame with 3 rows and 3 variables:
\describe{
  \item{group}{Different types of adverse events:
    "vascular disease", "GI" and "others" }
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
  \item{condition}{"vascular disease" and "GI" is conditioned on ICD-10 code
    originating from main diagnosis (variable "hdia")}
  \item{shar}{weights as defined by the Swedish Hip Arthroplasty Register}
  \item{sos}{weights as defined by
    The (Swedish) National Board of Health and Welfare "Socialstyrelsen"}
}}
\usage{
hip_adverse_events_icd10_old
}
\description{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\seealso{
Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{knee_adverse_events_icd10}},
  \code{\link{rxriskv}},
  \code{\link{tha_fracture_ae_icd10}},
  \code{\link{tha_fracture_ae_kva}}
}
\keyword{datasets}
