% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{hip_ae_hailer}
\alias{hip_ae_hailer}
\title{Classcodes for infection and dislocation after hip arthroplasty based on ICD-10
and KVA codes}
\format{
Data frame with 3 columns:
\describe{
  \item{group}{Infection or dislocation}
  \item{regex_icd10}{regular expressions based on ICD-10}
 \item{regex_kva}{regular expressions based on KVA codes}
}
}
\usage{
hip_ae_hailer
}
\description{
\code{\link{classcodes}} objects to use with \code{\link{classify}}.
}
\seealso{
ae

Other default classcodes: 
\code{\link{ae}},
\code{\link{charlson}},
\code{\link{cps}},
\code{\link{elixhauser}},
\code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
