% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{hip_fracture_ae_icd10}
\alias{hip_fracture_ae_icd10}
\title{Classcodes for adverse events after THA after hip fracture
based on ICD-10 codes}
\format{A data frame with 1 row and 2 variables:
\describe{
  \item{group}{We only have one group of adverse events}
  \item{regex}{regular expression identifying ICD-10 codes}
}}
\usage{
hip_fracture_ae_icd10
}
\description{
This classcodes object was developed by the Swedish Hip Arthroplasty Register.
The work was initiated for fracture data but the result might be feasible
also for non fracture data.
}
\details{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\references{
Codes partly inspired by p. 83 of the annual report 2016 from knee
arthroplasty register \url{http://www.myknee.se/pdf/SVK-2016_1.1.pdf}
}
\seealso{
Companion codes based on Atgardskoder (KVA codes)
\code{\link{hip_fracture_ae_kva}}.

Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{hip_fracture_ae_kva}},
  \code{\link{knee.ae_icd10}}, \code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
