% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_icd_data.R
\name{icd_data}
\alias{as.icd_data}
\alias{icd_data}
\alias{is.icd_data}
\title{Patient ICD codes}
\usage{
as.icd_data(x, ...)

is.icd_data(x)
}
\arguments{
\item{x}{object}

\item{...}{possible additional objects to merge with \code{x}}
}
\value{
\code{as.icd_data} returns an object of class \code{tbl_df} with columns:
\describe{
\item{lpnr}{patient id}
\item{date}{date when the icd code were valid}
\item{icd}{icd code}
}
\code{as.icd_data} reuturns \code{TRUE} if these same condityions are met,
\code{FALSE} otherwise.
(Note that \code{icd_data} is not a formal class of its own!)
}
\description{
Patient ICD codes
}
\examples{
\dontrun{
as.icd_data(oppen, sluten)
}
}

