% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_fun.R
\name{index_fun}
\alias{index_fun}
\alias{index}
\alias{index.data.frame}
\alias{index.matrix}
\title{Calculate index based on classification scheme}
\usage{
index(classified, ...)

\method{index}{data.frame}(classified, ...)

\method{index}{matrix}(classified, index = NULL, cc = NULL, ...)
}
\arguments{
\item{classified}{output from \code{\link[=classify]{classify()}}}

\item{...}{used internally}

\item{index}{name of column with 'weights' from corresponding
\code{\link{classcodes}} object. Can be \code{NULL} if the index is just a unweighted
count of all identified groups.}

\item{cc}{\code{\link{classcodes}} object. Can be \code{NULL} if a \code{classcodes} object is
already available as an attribute of \code{classified} (which is often the case)
and/or if \code{index = NULL}.}
}
\value{
Named numeric index vector with names corresponding to
\code{rownames(classified)}
}
\description{
This is the third step of \code{codify() \%>\% classify() \%>\% index()}.
The function takes classified case data and calculates
(weighted) index sums as specified by weights from a \code{classcodes} object.
}
\details{
Index weights for subordinate hierarchical classes
(as identified by \code{attr(cc, "hierarchy")}) are excluded in presence of
superior classes if index specified with argument \code{index}.
}
\seealso{
Other verbs: 
\code{\link{categorize}()},
\code{\link{classify}()},
\code{\link{codify}()}
}
\concept{verbs}
