% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{knee.ae_icd10}
\alias{knee.ae_icd10}
\alias{hip.ae_icd10}
\alias{hip.ae_kva}
\alias{knee.ae_kva}
\title{Classcodes for adverse events after knee and hip arthroplasty based on ICD-10 and KVA codes}
\format{Data frame with 3 columns:
\describe{
  \item{group}{Different types of adverse events (see reference section)}
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
  \item{condition}{two special conditions are used, see below.}
}}
\usage{
knee.ae_icd10
}
\description{
\code{\link{classcodes}} objects to use with \code{\link{classify}}
}
\details{
Group names are prefixed by two letters as given by the reference.
Two groups (DB and DM) are split into two due to different conditions.
}
\section{Conditions}{

Special conditions are used (see the reference):

\describe{
 \item{hbdia1_hdia}{\code{TRUE} if the code was
 given as any type of diagnose at first visit after TKA/THA, or as main diagnose
 for later visits, otherwise \code{FALSE}}
 \item{late_hdia}{\code{TRUE} if the code was
 given as main diagnose at a later visit after TKA/THA, otherwise \code{FALSE}}
 \item{post_op}{\code{TRUE} if the code was
 given at a later visit after TKA/THA, otherwise \code{FALSE}}
}
}

\references{
Codes for knee taken from p. 83 of the annual report 2016 from knee
arthroplasty register \url{http://www.myknee.se/pdf/SVK-2016_1.1.pdf}.
\code{hip.ae_icd10} is a modified version for hips based on
\code{knee.ae_icd10}.
}
\seealso{
Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{hip_fracture_ae_icd10}},
  \code{\link{hip_fracture_ae_kva}}, \code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
