% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{knee_adverse_events_icd10}
\alias{knee_adverse_events_icd10}
\alias{hip_adverse_events_icd10}
\title{Classcodes for adverse events after knee and hip arthroplasty based on ICD-10 codes}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{group}{Different types of adverse events (see reference section)}
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
  \item{condition}{two special conditions are used, see below.}
}}
\usage{
knee_adverse_events_icd10
}
\description{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\details{
Group names are prefixed by two letters as given by the reference.
Two groups (DB and DM) are split into two due to different conditions.
}
\section{Conditions}{

Two special conditions are used (see the reference):

\describe{
 \item{hbdia1_hdia}{a boolean variable with value \code{TRUE} if the code was
 given as any type of diagnose at first visit after TKA, or as main diagnose
 for later visits, otherwise \code{FALSE}}
 \item{late_hdia}{a boolean variable with value \code{TRUE} if the code was
 given as main diagnose at a later visit after TKA, otherwise \code{FALSE}}
}
}

\references{
Codes for knee taken from p. 83 of the annual report 2016 from knee
arthroplasty register \url{http://www.myknee.se/pdf/SVK-2016_1.1.pdf}.
\code{hip_adverse_events_icd10} is a modified version for hips based on
\code{knee_adverse_events_icd10}.
}
\seealso{
\code{\link{hip_adverse_events_icd10_old}} for a now utdated version
of hip adverse events.

Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{hip_adverse_events_icd10_old}},
  \code{\link{rxriskv}},
  \code{\link{tha_fracture_ae_icd10}},
  \code{\link{tha_fracture_ae_kva}}
}
\concept{default classcodes}
\keyword{datasets}
