% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr2codedata.R
\name{npr2codedata}
\alias{npr2codedata}
\title{Use data from the Swedish National Patient Register as codedata}
\usage{
npr2codedata(x, y = NULL, nprdate = "utdatuma", .copy = NA, ...)
}
\arguments{
\item{x, y}{data frames with columns specified below.}

\item{nprdate}{Name (as character) of date column to be recognized as code
date for out-patients: "indatuma" or "utdatuma".}

\item{.copy}{Should the object be copied internally by \code{\link{copy}}?
\code{NA} (by default) means that objects smaller than 1 GB are copied.
If the size is larger, the argument must be set explicitly. Set \code{TRUE}
to make copies regardless of object size. This is recommended if enough RAM
is available. If set to \code{FALSE}, calculations might be carried out
but the object will be changed by reference.}

\item{...}{arguments passed to \code{\link{as.codedata}}}
}
\value{
Output from \code{\link{as.codedata}}
}
\description{
Data from the Swedish National Patient Register (NPR) can be recognized as such if
\code{x} contains columns named \code{lpnr, indatuma/utdatuma} AND
(\code{hdia, bdia1, ..., bdia15} OR \code{op1, ..., op30}).
}
\details{
Note that the name should be \code{indatuma/utdatuma} (with letter "a"
suffixed). This is to distinguish the character date format \code{\%Y\%m\%d}
from the SPSS date format sometimes stored in \code{indatum/utdatum}.
If the SPSS date is the only one available, the help function
\code{\link{spss2date}} can be used for transformation, but this must be done
before using \code{as.codedata}.

If the data set is of this format, it is transformed to the codedata format
and returned as such.
Also, NPR data often comes in the form of two separate data sets, one for
inpatient and one for outpatient care. It is therefore possible to supply
an additional data set by argument \code{y}. It is assumed that data from
outpatient care has relevant dates stored as \code{utdatuma} and in-patient
data has relevant dates named \code{utdatuma}. This is however not tested
formally. If \code{utdatuma} is missing from the out-patient data,
\code{indatuma} is used without warning.
}
