% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{pardata}
\alias{pardata}
\alias{as.pardata}
\title{Data from the Swedish Patient Register (PAR).}
\usage{
as.pardata(x, y = NULL)
}
\arguments{
\item{x, y}{objects with data from NPR
(possibly one object for in- and one for outpatient data)}
}
\value{
Object as described by \code{\link{as.codedata}}
}
\description{
Functions to check that a data set contains patient data from
the Swedish Patient Register (PAR).
}
