% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcodes.R
\name{print.classcodes}
\alias{print.classcodes}
\title{Print classcodes object}
\usage{
\method{print}{classcodes}(x, n = NULL, ...)
}
\arguments{
\item{x}{object of type classcodes}

\item{n}{number of rows to preview (\code{n = 0} is allowed)}

\item{...}{arguments passed to print method for tibble}
}
\description{
Print classcodes object
}
\examples{
# Default printing
coder::elixhauser

# Print all rows
print(coder::elixhauser, n = 31)
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebook}()},
\code{\link{print.classified}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
