% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{rxriskv}
\alias{rxriskv}
\alias{rxriskv_icd10}
\alias{rxriskv_modified_icd10}
\title{Classcodes for RxRiskV (original and modified)}
\format{Data frames with 39 rows and 2 variables:
\describe{
  \item{group}{desease group}
  \item{regex}{regular expressions identifying ATC codes of each group}
}}
\usage{
rxriskv_icd10

rxriskv_modified_icd10
}
\description{
\code{\link{classcodes}} object to use with \code{\link{classify}}
}
\keyword{datasets}

