% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_classcodes.R
\name{set_classcodes}
\alias{set_classcodes}
\title{Set classcodes object}
\usage{
set_classcodes(
  cc,
  classified = NULL,
  regex = "regex",
  start = TRUE,
  stop = FALSE,
  tech_names = FALSE
)
}
\arguments{
\item{cc}{\code{\link{classcodes}} object or name of such}

\item{classified}{object that classcodes could be inherited from}

\item{regex}{name of column with regular expressions to use for
classification, either with or without prefix \code{regex_}}

\item{start, stop}{should codes start/end with the specified regular expressions?
If \code{TRUE}, column "regex" is prefixed/suffixed by "^"/"$".}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{by}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.}
}
\value{
\code{\link{classcodes}} object.
}
\description{
Set classcodes object
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{print.summary.classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
