% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Visualise classification scheme in web browser}
\usage{
show(x, group = NULL, show = TRUE)
}
\arguments{
\item{x}{classcodes specification as either a name or a classcodes object,
or a classcodes object itself.}

\item{group}{names (as character vector) of groups to visualise
(all groups if \code{NULL})}

\item{show}{should a visualisation be shown in a web browser.
Set to \code{FALSE} to just retrieve a URL for later use.}
}
\value{
URL to website with visualisation (invisible)
}
\description{
Classes are visualised by their regular expressions in a web browser.
The visualisation does not give any details on group names, conditions or
weights but might be useful both for understanding of a clasification scheme
in use, and during the creation and debugging of such.
}
\examples{
\dontrun{

 # How is depression classified according to Elixhauser?
 show("elix_icd10", "depression")

 # Compare the two diabetes groups according to Charlson
 show("charlson_icd10",
   c("diabetes without complication", "diabetes complication"))

 # Show all groups from Charlson
 show("charlson_icd10")

 # Get URL for later visualisation
 (show("tha_fracture_ae_icd10", show = FALSE))

}
}

