% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{tha_fracture_ae_kva}
\alias{tha_fracture_ae_kva}
\title{Classcodes for adverse events after THA after hip fracture
based on Atgardskoder (KVA codes)}
\format{A data frame with 1 row and 2 variables:
\describe{
  \item{group}{We only have one group of adverse events}
  \item{regex}{regular expression identifying KVA codes}
}}
\usage{
tha_fracture_ae_kva
}
\description{
This classcodes object was developed by the Swedish Hip Arthroplasty Register.
The work was initiated for fracture data but the result might be feasible
also for non fracture data.
}
\details{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\seealso{
Companion codes based on ICD-10 \code{\link{tha_fracture_ae_icd10}}.

Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{hip_adverse_events_icd10_old}},
  \code{\link{knee_adverse_events_icd10}},
  \code{\link{rxriskv}},
  \code{\link{tha_fracture_ae_icd10}}
}
\concept{default classcodes}
\keyword{datasets}
