% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.classcodes}
\alias{visualize.classcodes}
\title{Visualize classification scheme in web browser}
\usage{
visualize.classcodes(x, group = NULL, show = TRUE, ...)
}
\arguments{
\item{x}{\link{classcodes} object or name of such object included in the package
(see \code{\link[=all_classcodes]{all_classcodes()}}).}

\item{group}{names (as character vector) of groups to visualize
(subset of \code{rownames(x)}). (All groups if \code{NULL}.)}

\item{show}{should a visualization be shown in the default web browser.
Set to \code{FALSE} to just retrieve a URL for later use.}

\item{...}{
  Arguments passed on to \code{\link[=set_classcodes]{set_classcodes}}
  \describe{
    \item{\code{regex}}{name of column with regular expressions to use for
classification.
\code{NULL} (default) uses \code{attr(obj, "regexpr")[1]}.}
  }}
}
\value{
URL to website with visualization (invisible)
}
\description{
Groups from a \code{classcodes} object are visualized by their regular expressions
in the default web browser.
The visualization does not give any details on group names, conditions or
weights but might be useful both for understanding of a classification scheme
in use, and during the creation and debugging of such.
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebook}()},
\code{\link{print.classcodes}()},
\code{\link{print.classified}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()}
}
\concept{classcodes}
