% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{ae}
\alias{ae}
\alias{knee.ae_icd10}
\alias{hip.ae_icd10}
\alias{hip.ae_kva}
\alias{knee.ae_kva}
\alias{hip.fracture.ae_icd10}
\title{Classcodes for adverse events after knee and hip arthroplasty based on ICD-10
and KVA codes}
\format{Data frame with 3 columns:
\describe{
  \item{group}{Different types of adverse events (see reference section)}
  \item{regex}{regular expressions identifying ICD-10 or KVA codes for each group}
  \item{condition}{special conditions are used, see below.}
}}
\usage{
knee.ae_icd10

hip.ae_icd10

hip.ae_kva

knee.ae_kva

hip.fracture.ae_icd10
}
\description{
\code{\link{classcodes}} objects to use with \code{\link{classify}}
}
\details{
Group names are prefixed by two letters as given by the references.
Two groups (DB and DM) are split into two due to different conditions.
}
\section{Hip fractures}{

AE codes for hip fractures are based on codes for elective cases but with
some additional codes for DM 1 (N300, N308, N309 and N390).
}

\section{Conditions}{

Special conditions apply to all categories.
Those require non-standard modifications
of the classcodes data prior to categorization.

\describe{
 \item{hbdia1_hdia}{\code{TRUE} if the code was
 given as any type of diagnose during hospital visit for index operation,
 or as main diagnose for later visits, otherwise \code{FALSE}}
 \item{late_hdia}{\code{TRUE} if the code was
 given as main diagnose at a later visit after the index operation,
 otherwise \code{FALSE}}
 \item{post_op}{\code{TRUE} if the code was
 given at a later visit after the index operation, otherwise \code{FALSE}}
}
}

\references{
Knee (p. 83): \url{http://www.myknee.se/pdf/SVK-2016_1.1.pdf}.
Hip (p. 149): \url{https://registercentrum.blob.core.windows.net/shpr/r/-rsrapport-2017-S1xKMzsAwX.pdf}
}
\seealso{
Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
