% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classcodes.R
\name{all_classcodes}
\alias{all_classcodes}
\title{Summary data for all default classcodes object in the package}
\usage{
all_classcodes()
}
\value{
Data frame with columns describing all default classcodes
objects from the package.
}
\description{
Summary data for all default classcodes object in the package
}
\examples{
all_classcodes()
}
\seealso{
Other classcodes: 
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
