% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{as.codedata}
\alias{as.codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ..., .setkeys = TRUE, alnum = FALSE)

is.codedata(x)
}
\arguments{
\item{x}{\code{\link{data.frame}} with columns "id" (character), "code",
and optional "code_date" (Date).}

\item{...}{
  Arguments passed on to \code{\link[=dates_within]{dates_within}}
  \describe{
    \item{\code{from}}{first and last date of interval to compare with}
    \item{\code{to}}{first and last date of interval to compare with}
  }}

\item{.setkeys}{should output be indexed (Beneficial for large data sets.)}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters?}
}
\value{
\code{\link{data.table}} with columns: "id" (character), "code" and
  optionally "code_date" (Date)
  Possible additional columns from \code{x} are preserved.
}
\description{
Make structured code data set based on \code{\link{data.frame}}.
}
\examples{

x <- data.frame(
  id = "1",
  code_date = as.Date("2017-02-02"),
  code = "a",
  stringsAsFactors = FALSE
)
as.codedata(x)

# Drop dates outside specified limits
y <- data.frame(id = "2", code_date = as.Date("3017-02-02"), code = "b",
  stringsAsFactors = FALSE)
z <- rbind(x, y)
as.codedata(z)

}
\seealso{
npr2codedata
}
\concept{codedata}
