% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{as.codedata}
\alias{as.codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ..., .setkeys = TRUE, alnum = FALSE)

is.codedata(x)
}
\arguments{
\item{x}{data frame with columns "id" (character), "code",
and optional "code_date" (Date).}

\item{...}{by default, codes for future dates or dates before "1970-01-01"
are ignored (with a warning). Set \code{to, from}
(passed to \code{\link{filter_dates}}) to override.}

\item{.setkeys}{should keys be set for the \code{\link{data.table}} object?
(Beneficial for large data sets.)}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters?}
}
\value{
\code{\link{data.table}} with columns: "id" (character), "code" and
  optionally "code_date" (Date)
  Possible additional columns from \code{x} are preserved.
}
\description{
Code data
}
\examples{

x <- data.frame(
  id = 1,
  code_date = as.Date("2017-02-02"),
  code = "a"
)
as.codedata(x)

# Drop dates outside specified limits
y <- data.frame(id = 2, code_date = as.Date("3017-02-02"), code = "b")
z <- rbind(x, y)
as.codedata(z)

}
\seealso{
npr2codedata
}
\concept{codedata}
