% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{as.codedata}
\alias{as.codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ..., .setkeys = TRUE, alnum = FALSE)

is.codedata(x)
}
\arguments{
\item{x}{\link{data.frame} with columns \code{id} (character), \code{code},
and optional \code{code_date} (Date).}

\item{...}{
  Arguments passed on to \code{\link[=dates_within]{dates_within}}
  \describe{
    \item{\code{from}}{first date of interval to compare with (\code{1970-01-01} by default,
since we assume that little data was recorded before 'epoch'; this is a bold
assumption, however, which should be explicitly re-considered for relevant
applications)}
    \item{\code{to}}{last date of interval to compare with
(the current data from \code{\link[=Sys.Date]{Sys.Date()}}, by default,
since we assume that future events are unknown,
thus indicating coding errors)}
  }}

\item{.setkeys}{should output be indexed (Beneficial for large data sets.)}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters?}
}
\value{
\link{data.table} with columns: \code{id} (character), \code{code} and
optionally \code{code_date} (Date)
Possible additional columns from \code{x} are preserved.
}
\description{
Make structured code data set based on \link{data.frame}.
}
\examples{

x <-
  data.frame(
    id               = "1",
    code_date        = as.Date("2017-02-02"),
    code             = "a",
    stringsAsFactors = FALSE
  )
as.codedata(x)

# Drop dates outside specified limits
y <-
  data.frame(
    id               = "2",
    code_date        = as.Date("3017-02-02"),
    code             = "b",
    stringsAsFactors = FALSE
  )
z <- rbind(x, y)
as.codedata(z)

}
\concept{codedata}
