% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{as.codedata}
\alias{as.codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ..., .setkeys = TRUE, alnum = FALSE)

is.codedata(x)
}
\arguments{
\item{x}{data frame with columns "id", "code_date" and "code"}

\item{...}{by default, codes for future dates or dates before "1970-01-01"
are ignored (with a warning). Specify \code{to, from}
(as passed to \code{\link{filter_dates}}) to override.}

\item{.setkeys}{should keys be set for the data.table object?
To initially set the key for a large data set can be slow
but it could be beneficial for later use of the object.}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters (Boolean)?}
}
\value{
\code{as.codedata} returns a \code{\link{data.table}} object
  with mandatory columns: "id" (individual id coerced to character) and
  "code". It has an optional column
  "code_date" (when the code was valid) if there was such a column in
  \code{x}.
  Possible additional columns from \code{x} are preserved.

\code{is.codedata} returns \code{TRUE} if these same conditions are met and
\code{FALSE} otherwise.
(Note that \code{codedata} is not a formal R class of its own!)
}
\description{
Code data
}
\examples{

x <- data.frame(
  id = 1,
  code_date = as.Date("2017-02-02"),
  code = "a"
)
as.codedata(x)

# Drop dates outside specified limits
y <- data.frame(id = 2, code_date = as.Date("3017-02-02"), code = "b")
z <- rbind(x, y)
as.codedata(z)

}
\seealso{
npr
}
\concept{codedata}
