% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.classified.R
\name{as.data.frame.classified}
\alias{as.data.frame.classified}
\title{Convert output from classify to data frame}
\usage{
\method{as.data.frame}{classified}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{classify}}
(technically an object of internal class "classified")}

\item{...}{ignored}
}
\value{
data frame with:
\itemize{
  \item{first column named as "id" column specified as input
    to \code{\link{classify}} and with data from \code{row.names(x)}}
  \item{all columns from \code{x}}
  \item{no row names}
}
}
\description{
Output from \code{\link{classify}} is given as a matrix.
It is often convinient co convert this matrix to a data frame with
a separate column for id instead of row names since this fits better with
the philospohy of "tidyverse". The data frame format is however less optimal
for working wit hlarge data sets, wherefore matrix output is used as default.
}

