% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{as.data.frame.classified}
\alias{as.data.frame.classified}
\alias{as.data.table.classified}
\title{Convert output matrix from classify to data frame or data.table}
\usage{
\method{as.data.frame}{classified}(x, ...)

\method{as.data.table}{classified}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{classify}}}

\item{...}{ignored}
}
\value{
data frame with:
\itemize{
  \item{first column named as "id" column specified as input
    to \code{\link{classify}} and with data from \code{row.names(x)}}
  \item{all columns from \code{classified}}
  \item{no row names}
}
}
\description{
Convert output matrix from classify to data frame or data.table
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
