% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.keyvalue.R
\name{as.keyvalue.classcodes}
\alias{as.keyvalue.classcodes}
\title{Make keyvalue object from classcodes object}
\usage{
\method{as.keyvalue}{classcodes}(x, coding, ...)
}
\arguments{
\item{x}{classcodes object}

\item{coding}{either a vector with codes from the original classification,
or a name (character vector of length one) keyvalue object from package
"decoder" (for example "icd10" or "atc")}

\item{...}{arguments passed to \code{\link{summary.classcodes}}}
}
\description{
Make keyvalue object from classcodes object
}
