% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize cases based on external data and classification scheme}
\usage{
categorize(
  .data,
  from,
  what,
  id,
  ind = NULL,
  sort = TRUE,
  codify_args = list(),
  cc_args = list()
)
}
\arguments{
\item{.data}{To which data set should the categories be added?
A data frame with at least one column with case (patient) identification
(column name specified by argument id). A column with a date of interest
(column name specified by argument date) is mandatory if days != NULL.}

\item{from}{where do we find the class codes? \code{\link{as.codedata}}}

\item{what}{What to add? Specify classification scheme
(\code{\link{classcodes}} classcode object) specifying categories to add
based on individual code groups.}

\item{id}{name (as character) of variable in \code{x} to group id (for
example a patient id).}

\item{ind}{control possible inclusion of index vector. Set to \code{FALSE} if
no index should be calculated, otherwise a value passed to argument
\code{by} in function \code{\link{index}}. It is possible to
include several indices as a character vector.
\code{NULL} will include all available indices.}

\item{sort}{logical. Should output be sorted by the 'id' column?
(This could effect computational speed for large data sets.)
Data is sorted by 'id' internally. It is therefore faster to keep the output
sorted this way, but this might be inconvinient if the original
order was intended. Set to \code{FALSE} in order to not shuffle the
input data.}

\item{codify_args}{List of named arguments passed to \code{\link{codify}}}

\item{cc_args}{List with named arguemnts passed to \code{\link{set_classcodes}}}
}
\value{
data frame ('data.table' object) with all data from
the input data set \code{to/.data} combined with logical columns indicating
membership of categories identified by the classcode object.
Indices are also included if so specified by the 'ind' argument.
}
\description{
\code{categorize} is the main function of the package.
\code{categorise} is used as an additional alias.
}
\examples{
# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, ex_icd10, "elixhauser", id = "name")

# Add Charlson categorias and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, ex_icd10, "charlson",
  id = "name",
  ind = c("quan_original", "quan_updated"),
  codify_args = list(date = "surgery", days = c(-30, -1)),
  cc_args = list(tech_names = TRUE)
)
}
\seealso{
Other verbs: 
\code{\link{classify}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
