% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize cases based on external data and classification scheme}
\usage{
categorize(
  .data,
  from,
  what,
  id,
  date = NULL,
  days = NULL,
  ind = NULL,
  tech_names = FALSE,
  sort = TRUE,
  regex = "regex",
  alnum = FALSE
)
}
\arguments{
\item{.data}{To which data set should the categories be added?
A data frame with at least one column with case (patient) identification
(column name specified by argument id). A column with a date of interest
(column name specified by argument date) is mandatory if days != NULL.}

\item{from}{where do we find the class codes? \code{\link{as.codedata}}}

\item{what}{What to add? Specify classification scheme
(\code{\link{classcodes}} classcode object) specifying categories to add
based on individual code groups.}

\item{id, date, days}{arguments passed to \code{\link{codify}}}

\item{ind}{control possible inclusion of index vector. Set to \code{FALSE} if
no index should be calculated, otherwise a value passed to argument
\code{by} in function \code{\link{index}}. It is possible to
include several indices as a character vector.
\code{NULL} will include all available indices.}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{by}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.}

\item{sort}{logical. Should output be sorted by the 'id' column?
(This could effect computational speed for large data sets.)
Data is sorted by 'id' internally. It is therefore faster to keep the output
sorted this way, but this might be inconvinient if the original
order was intended. Set to \code{FALSE} in order to not shuffle the
input data.}

\item{regex}{name of column with regular expressions from classcodes object
\code{what} to use for classification, either with or without prefix "regex_"}

\item{alnum}{argument passed to \code{\link{as.codedata}}.
Specify if all non alphanumeric characters should be dropped from codes.}
}
\value{
data frame ('data.table' object) with all data from
the input data set \code{to/.data} combined with logical columns indicating
membership of categories identified by the classcode object.
Indices are also included if so specified by the 'ind' argument.
}
\description{
\code{categorize} is the main function of the package.
\code{categorise} is used as an additional alias.
}
\examples{
# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, ex_icd10, "elixhauser", id = "name")

# Add Charlson categorias and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, ex_icd10, "charlson",
  id = "name", date = "surgery", days = c(-30, -1),
  ind = c("quan_original", "quan_updated"),
  tech_names = TRUE
)
}
\seealso{
Other verbs: 
\code{\link{classify}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
