% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\alias{categorize.data.frame}
\alias{categorize.tbl_df}
\alias{categorize.data.table}
\alias{categorize.codified}
\title{Categorize cases based on external data and classification scheme}
\usage{
categorize(x, ...)

\method{categorize}{data.frame}(x, ...)

\method{categorize}{tbl_df}(x, ...)

\method{categorize}{data.table}(x, ..., codedata, id, code, codify_args = list())

\method{categorize}{codified}(
  x,
  ...,
  cc,
  index = NULL,
  sort = TRUE,
  cc_args = list(),
  check.names = TRUE,
  .data_cols = NULL
)
}
\arguments{
\item{x}{data set with mandatory id column
(identified by argument \code{id}),
and optional column with date of interest
(identified by argument \code{date} if  \code{days != NULL}).
Alternatively, the output from \code{\link[=codify]{codify()}}}

\item{...}{arguments passed between methods}

\item{codedata}{external code data}

\item{id}{name of code/id columns (in \code{codified}).}

\item{code}{name of code/id columns (in \code{codified}).}

\item{cc}{\code{\link{classcodes}} object (or name of such object).}

\item{index}{A character vector of index values to calculate (passed to \code{\link[=index]{index()}}.
Set to \code{FALSE} if no index should be calculated.
If \code{NULL}, the default, all available indices (from \code{attr(cc, "indices")})
are provided. A message lists the indices so that you can check they're
correct; suppress the message by supplying \code{index} explicitly.}

\item{sort}{logical. Should output be sorted by the 'id' column?
(This could affect computational speed for large data sets.
Data is sorted by 'id' internally. It is therefore faster to keep the
output sorted this way, but this might be inconvenient if the original
order was important.)}

\item{cc_args, codify_args}{List of named arguments passed to
\code{\link[=set_classcodes]{set_classcodes()}} and  \code{\link[=codify]{codify()}}}

\item{check.names}{Column names are based on \code{cc$group}, which might include
spaces. Those names are changed to syntactically correct names by
\code{check.names = TRUE}. Syntactically invalid, but grammatically correct
names might be preferred for presentation of the data as achieved by
\code{check.names = FALSE}. Alternatively, if \code{categorize} is called repeatedly,
longer informative names might be created by \code{cc_args = list(tech_names = TRUE)}.}

\item{.data_cols}{used internally}
}
\value{
Object of the same class as \code{x} with additional logical columns
indicating membership of categories identified by the
\code{classcodes} object (the \code{cc} argument).
Indices are also included if specified by the 'index' argument.
}
\description{
This is the main function of the package, which relies of a triad of objects:
(1) \code{data} with unit id:s and possible dates of interest;
(2) \code{codedata} for corresponding
units and with optional dates of interest and;
(3) a classification scheme (\code{\link{classcodes}} object; \code{cc}) with regular
expressions to identify and categorize relevant codes.
}
\details{
The function combines the three underlying steps performed by
\code{\link[=codify]{codify()}}, \code{\link[=classify]{classify()}} and \code{\link[=index]{index()}}.
Relevant arguments are passed to those functions by
\code{codify_args} and \code{cc_args}.
index
}
\examples{
# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
  id = "name", code = "icd10")

# Add Charlson categories and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, codedata = ex_icd10, cc = "charlson",
  id = "name", code = "icd10",
  index = c("quan_original", "quan_updated"),
  codify_args =
    list(date = "surgery", days = c(-30, -1), code_date = "admission"),
  cc_args = list(tech_names = TRUE)
)
}
\seealso{
Other verbs: 
\code{\link{classify}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
