% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize cases based on external data and classcodes (classification scheme)}
\usage{
categorize(
  data,
  codedata,
  cc,
  id,
  index = NULL,
  sort = TRUE,
  codify_args = list(),
  cc_args = list()
)
}
\arguments{
\item{data}{(\code{\link{data.frame}}) with case data including at least an id,
and possibly a date, column.}

\item{codedata}{external code data (see \code{\link{as.codedata}})}

\item{cc}{\code{\link{classcodes}} object (or name of such).}

\item{id}{name of code/id columns (in \code{codified}).}

\item{index}{control possible inclusion of index vector.
Set to \code{FALSE} if no index should be calculated, otherwise a value passed to argument
\code{by} in function \code{\link{index}}. It is possible to
include several indices as a character vector.
\code{NULL} will include all available indices.}

\item{sort}{logical. Should output be sorted by the 'id' column?
(Sorting by id is always done internally. To restore the original row order of \code{data}
could be slow for large data sets.)}

\item{codify_args}{List of named arguments passed to \code{\link{codify}}}

\item{cc_args}{List with named arguments passed to \code{\link{set_classcodes}}}
}
\value{
\code{\link{data.table}} with
 \code{data} enhanced by logical columns indicating
membership of categories identified by \code{cc}, as well as
indices if specified by \code{index}.
}
\description{
Categorize cases based on external data and classcodes (classification scheme)
}
\examples{
# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, ex_icd10, "elixhauser", id = "name")

# Add Charlson categorias and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, ex_icd10, "charlson",
  id = "name",
  index = c("quan_original", "quan_updated"),
  codify_args = list(date = "event", days = c(-30, -1)),
  cc_args = list(tech_names = TRUE)
)
}
\seealso{
Other verbs: 
\code{\link{classify}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
