% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize cases based on external data and classification scheme}
\usage{
categorize(
  data,
  codedata,
  cc,
  id,
  index = NULL,
  sort = TRUE,
  codify_args = list(),
  cc_args = list()
)
}
\arguments{
\item{data}{data frame with mandatory id column
(identified by argument \code{id}),
and semi-optional column with date of interest
(identified by argument \code{date} if \code{days != NULL}).}

\item{codedata}{external code data (see \code{\link{as.codedata}})}

\item{cc}{\code{\link{classcodes}} object (or name of such object).}

\item{id}{name of code/id columns (in \code{codified}).}

\item{index}{control possible inclusion of index vector.
Set to \code{FALSE} if no index should be calculated, otherwise a value
passed to argument
\code{by} in function \code{\link{index}}. It is possible to
include several indices as a character vector.
\code{NULL} will include all available indices as specified by the
"indices" attribute of the \code{cc} object (\code{attr(cc, "indices")})}

\item{sort}{logical. Should output be sorted by the 'id' column?
(This could effect computational speed for large data sets.)
Data is sorted by 'id' internally. It is therefore faster to keep the
output sorted this way, but this might be inconvenient if the original
order was intended. Set to \code{FALSE} in order to not shuffle the
input data.}

\item{codify_args}{List of named arguments passed to \code{\link{codify}}}

\item{cc_args}{List with named arguments passed to
\code{\link{set_classcodes}}}
}
\value{
Object of class \code{data.table} made from \code{data} combined with
logical columns indicating membership of categories identified by the
\code{classcodes} object (the \code{cc} argument).
Indices are also included if specified by the 'index' argument.
}
\description{
This is the main function of the package, which relies of a triad of objects:
(1) \code{data}) with unit id:s and possible dates of interest;
(2) \code{codedata} for corresponding
units and with optional dates of interest and;
(3) a classification scheme ('classcodes' object; \code{cc}) with regular
expressions to identify and categorize relevant codes.
}
\details{
The function combines the three underlying steps performed by
\code{\link{codify}}, \code{\link{classify}} and \code{\link{index_fun}}.
 Relevant arguments are passed forward to those functions by
 \code{codify_args} and \code{cc_args}.
}
\examples{
# Add Elixhauser based on all registered ICD10-codes
categorize(ex_people, ex_icd10, "elixhauser", id = "name")

# Add Charlson categories and two versions of a calculated index.
# Only include recent hospital visits within 30 days before surgery,
# and use technical variable names to clearly identify the new columns.
categorize(ex_people, ex_icd10, "charlson",
  id = "name",
  index = c("quan_original", "quan_updated"),
  codify_args = list(date = "event", days = c(-30, -1)),
  cc_args = list(tech_names = TRUE)
)
}
\seealso{
For more details see the help vignette:
\code{vignette("coder", package = "coder")} and the package website
\url{https://eribul.github.io/coder}

Other verbs: 
\code{\link{classify}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
