% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcodes.R
\name{classcodes}
\alias{classcodes}
\alias{as.classcodes}
\alias{is.classcodes}
\title{Classcodes methods}
\usage{
as.classcodes(x, hierarchy = attr(x, "hierarchy"), .name = NULL)

is.classcodes(x)
}
\arguments{
\item{x}{data frame with properties as described in the details section}

\item{hierarchy}{list of pairwise group names to appear as superior and
subordinate classes.
  To be used for indexing when the subordinate class is redundant.}

\item{.name}{used internally for name dispatch}
}
\value{
Object of class "classcodes" (data frame) with attribute \code{code}
specifying the coding used (for example "icd10", or "ATC").
Could be \code{NULL} for unknown or arbitrary coding.
}
\description{
"Classcodes" are classifying schemes based on regular expression stored in
data frames. These are essential to the package and constitute the third
part of the triad of case data, code data and classification scheme.
}
\details{
A classcodes object is a data frame with columns:
\describe{
\item{group:}{unique and non missing class name}
\item{regex:}{regular expression defining class membership
(see \code{\link{regex}} for details). Occurrences of non unique regular
expressions will lead to the same class having multiple names. This is
accepted but will raise a warning.}
\item{description (optional):}{description of each category}
\item{condition (optional):}{a class might have conditions additional to what
  is expressed by \code{regex}. If so, these should be specified as quoted
  expressions that can be evaluated within the data frame used by
  \code{\link{classify}}}
\item{weights (optional):}{weights for each class used for
  \code{\link{index}} calculation.
  Could be more than one and could have arbitrary names.}
  \item{hierarchy}{pairwise hierarchys for wich only the superior class
    should be used for indexing}
}
Note that classes does not have to be disjunct.

The package have several default classcodes included listed on its index page
`help(package = "coder")`
}
\examples{
as.classcodes(coder::elixhauser)
}
\seealso{
Regular expressions used in classcodes can often be quite complex
and hard to interpret. Use \code{\link{visualize}} to visualize classcodes
graphically or \code{\link{summary.classcodes}} to
list all codes individually.

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
