% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcodes.R
\name{classcodes}
\alias{as.classcodes}
\alias{classcodes}
\alias{is.classcodes}
\title{Classcodes methods}
\usage{
as.classcodes(x)

is.classcodes(x)
}
\arguments{
\item{x}{data frame with properties as described in the details section}
}
\value{
Object of class "classcodes"
}
\description{
"Classcodes" are classifying schemes based on regular expression stored in
data frames. These are essential to the package and constitute the third
part of the triad of case data, code data and classification scheme.
}
\details{
A classcodes object is a data frame with columns:
\describe{
\item{group:}{unique and non missing class name}
\item{regex:}{regular expression defining class membership
(see \code{\link{regex}} for details). Occurrences of non unique regular
expressions will lead to the same class having multiple names. This is
accepted but will raise a warning.}
\item{condition (optional):}{a class might have conditions additional to what
  is expressed by \code{regex}. If so, these should be specified as quoted
  expressions that can be evaluated within the data frame used by
  \code{\link{classify}}}
\item{weights (optional):}{weights for each class used for
  \code{\link{index}} calculation.
  Could be more than one and could have arbitrary names.}
}
Note that classes does not have to be disjunct.

The package have several default classcodes included listed on its index page
`help(package = "classifyr")`
}
\examples{
as.classcodes(classifyr::elix_icd10)
}
\seealso{
Regular expressions used in classcodes can often be quite complex
and hard to interpret. Use \code{\link{show}} to visualize classcodes
graphically.
}

