% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcodes.R
\name{classcodes}
\alias{as.classcodes}
\alias{classcodes}
\alias{is.classcodes}
\title{Classcode methods}
\usage{
as.classcodes(x)

is.classcodes(x)
}
\arguments{
\item{x}{data.frame with properties as described in the details section}
}
\value{
Object of class "classcodes"
}
\description{
"Classcodes" are classifying schemes based on regular expression.
}
\details{
A classcode object have columns:
\describe{
\item{group:}{unique and non empty (NA or "" if character) group name}
\item{regex:}{unique regular expression defining class membership}
\item{w (optional):}{wieights for each class used for \code{\link{index}}
calculation}
}
Note that classes does not have to be disjunct.
}
\examples{
as.classcodes(classifyr::elix_icd10)
}

