% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify.default}
\alias{classify.codified}
\alias{classify.data.frame}
\alias{classify.data.table}
\title{Classify cases based on classcodes}
\usage{
classify(codified, cc, ..., cc_args = list())

\method{classify}{default}(codified, cc, ..., cc_args = list())

\method{classify}{codified}(codified, ...)

\method{classify}{data.frame}(codified, ...)

\method{classify}{data.table}(codified, cc, ..., id, code, cc_args = list())
}
\arguments{
\item{codified}{output from \code{\link[=codify]{codify()}}}

\item{cc}{\code{\link{classcodes}} object or name of such}

\item{...}{arguments passed between methods}

\item{cc_args}{List with named arguments passed to
\code{\link[=set_classcodes]{set_classcodes()}}}

\item{code, id}{name of code/id columns (in \code{codified}).}
}
\value{
Boolean matrix with one row for each element/row of \code{codified}
and columns for each class with corresponding class names (according to the
\code{\link{classcodes}} object).
}
\description{
Classify cases based on classcodes
}
\examples{
# Classify individual ICD10-codes by Elixhauser
classify(c("C80", "I20", "unvalid_code"), "elixhauser")

# Classify patients by Charlson for comorbidities during
# one year before surgery
x <- codify(ex_people, ex_icd10, id = "name",
  date = "event", days = c(-365, 0))
y <- classify(x, "charlson")

# Use tha RCS classification instead and use technical column names
y <- classify(x, "charlson",
  cc_args = list(regex = "icd10_rcs", tech_names = TRUE))

# It is possible to convert the output of classify to a data frame with
# id column instead of row names
as.data.frame(y)
}
\seealso{
\code{\link[=as.data.frame.classified]{as.data.frame.classified()}} for a convenience function to
convert the output of \code{\link[=classify]{classify()}} to a data frame with id column instead
of row names.

Other verbs: 
\code{\link{categorize}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
