% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify.default}
\alias{classify.codified}
\alias{classify.data.frame}
\alias{classify.data.table}
\title{Classify codified data}
\usage{
classify(codified, cc, ..., cc_args = list())

\method{classify}{default}(codified, cc, ..., cc_args = list())

\method{classify}{codified}(codified, ...)

\method{classify}{data.frame}(codified, ...)

\method{classify}{data.table}(codified, cc, ..., id, code, cc_args = list())
}
\arguments{
\item{codified}{output from \code{\link[=codify]{codify()}}}

\item{cc}{\code{\link{classcodes}} object (or name of a default object from
\code{\link[=all_classcodes]{all_classcodes()}}).}

\item{...}{arguments passed between methods}

\item{cc_args}{List with named arguments passed to
\code{\link[=set_classcodes]{set_classcodes()}}}

\item{code, id}{name of code/id columns (in \code{codified}).}
}
\value{
Object of class "classified". Inheriting from a Boolean matrix with
one row for each element/row of \code{codified}
and columns for each class with corresponding class names (according to the
\code{\link{classcodes}} object). Note, however, that \code{\link[=print.classified]{print.classified()}} preview
this output as a tibble.
}
\description{
This is the second step of \code{codify() \%>\% classify() \%>\% index()}.
Hence, the function takes a codified data set and classify each case based on
relevant codes as identified by the classification scheme provided by a
\code{classcodes} object.
}
\seealso{
\code{\link[=as.data.frame.classified]{as.data.frame.classified()}}, \code{\link[=as.data.table.classified]{as.data.table.classified()}} and
\code{\link[=as.matrix.classified]{as.matrix.classified()}}, \code{\link[=print.classified]{print.classified()}}

Other verbs: 
\code{\link{categorize}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
