% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify.default}
\alias{classify.data.frame}
\title{Classify elements}
\usage{
classify(x, by, ..., regex = "regex")

\method{classify}{default}(x, by, ..., regex = "regex")

\method{classify}{data.frame}(
  x,
  by,
  id = NULL,
  code = "code",
  tech_names = FALSE,
  ...,
  regex = "regex"
)
}
\arguments{
\item{x}{object with elements to classify (often the output from
\code{\link{codify}})}

\item{by}{classification scheme of type \code{classcodes} to classify by}

\item{...}{arguments passed to \code{\link{get_classcodes}}}

\item{regex}{name of column with regular expressions to use for
classification, either with or without prefix \code{regex_}}

\item{id}{name (as character) of variable in \code{x} to group id (for
example a patient id)}

\item{code}{name (as character) of variable in \code{x} containing codes to
classify}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{by}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.}
}
\value{
Boolean matrix with one row for each element/row of \code{x} and
  columns for each class with corresponding class names (according to the
  \code{\link{classcodes}} object).
}
\description{
This is the main functions of the package, performing the actual
classification of cases. It is often used with output from
\code{\link{codify}} as input and is sometimes followed by the
\code{\link{index}} function. These three function could easily be chained
together if using the magrittr pipe (\code{\%>\%}).
}
\details{
Note that row order is not preserved for \code{classify.data.frame} due to
 performance reasons (the function is intended to work with large data sets
 where sorting is computationally expensive).

 Row names does however identify origin (as specified by \code{id} for
 \code{classify.data.frame} or simply as \code{x} itself for
 \code{classify.default}. It should therefore be possible to manually
 reorder the output to maintain original order.
}
\examples{
# Classify individual ICD10-codes by Elixhauser
classify(c("C80", "I20", "unvalid_code"), by = "elixhauser")

# Classify patients by Charlson for comorbidities during
# one year before surgery
x <- codify(ex_people, ex_icd10, id = "name",
  date = "surgery", days = c(-365, 0))
(y <- classify(x, "charlson"))

# It is possible to convert the outpot of classify to a data frame with
# id column instead of row names
as.data.frame(y)
}
\seealso{
\code{\link{as.data.frame.classified}} for a convenience function to
convert the output of \code{classify} to a data frame with id column instead
of row names.

Other verbs: 
\code{\link{categorize}()},
\code{\link{codify}()},
\code{\link{index_fun}}
}
\concept{verbs}
