% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify.data.frame}
\alias{classify.default}
\title{Classify elements}
\usage{
classify(x, by, ...)

\method{classify}{default}(x, by, ...)

\method{classify}{data.frame}(x, by, id = NULL, code = "code",
  drop = FALSE, ...)
}
\arguments{
\item{x}{object with elements to classify}

\item{by}{classification scheme of type \code{classcodes} to classify by}

\item{id}{name (as character) of variable in \code{x} to
group id (for example a patient id)}

\item{code}{name (as character) of variable in \code{x} containing codes to classify}

\item{drop}{drop cases that does not belong to any class}
}
\value{
Boolean matrix with one row for each element/row of \code{x} and columns
for each class with corresponding class names (according to the
\code{\link{classcodes}} object).

Note that row order is
not preserved for \code{classify.data.frame}.

Row names identify origin (as specified id argument \code{id} for
\code{classify.data.frame} or simply as \code{x} itself for
\code{classify.default}.
}
\description{
Classify elements
}
\examples{
# Classify individual ICD10-codes by Elixhauser
classify(c("C80", "I20", "XXX"), by = "elix_icd10")

# Classify patients by Charlson
codify(ex_people, ex_icd10, id = "name", date = "surgery") \%>\%
classify("charlson_icd_10")
}

