% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{classify.data.frame}
\alias{classify.default}
\title{Classify elements}
\usage{
classify(x, ...)

\method{classify}{default}(x, classcodes = "elix_icd10", ...)

\method{classify}{data.frame}(x, classcodes = "elix_icd10", code = "icd",
  by = "lpnr", drop = FALSE, ...)
}
\arguments{
\item{x}{object with elements to classify}

\item{classcodes}{classification scheme of type \code{classcodes}}

\item{code}{name (as character) of variable in \code{x} containing codes to classify}

\item{by}{name (as character) of variable in \code{x} to
group by (for example a patient id)}

\item{drop}{drop cases that does not belong to any class}
}
\value{
Boolean matrix with one row for each element/row of \code{x} and columns
for each class with corresponding class names (according to the
\code{\link{classcodes}} object).

Note that row order is
not preserved for \code{classify.data.frame}.

Row names identify origin (as specified by argument \code{by} for
\code{classify.data.frame} or simply as \code{x} itself for
\code{classify.default}.
}
\description{
Classify elements
}
\examples{
# Classify patients using Elixhauser based on ICD10-codes
classify(c("C80", "I20", "XXX"))
}

