% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Make codebook for classcodes object}
\usage{
codebook(object, coding, ..., file = NULL)
}
\arguments{
\item{object}{classcodes object}

\item{coding}{either a vector with codes from the original classification,
or a name (character vector of length one) keyvalue object from package
"decoder" (for example "icd10cm" or "atc")}

\item{...}{
  Arguments passed on to \code{\link[=summary.classcodes]{summary.classcodes}}
  \describe{
    \item{\code{cc_args}}{List of named arguments passed to \code{\link[=set_classcodes]{set_classcodes()}}}
  }}

\item{file}{name of Excel file for data export}
}
\value{
List of data frames describing relationship
between groups and individual codes
}
\description{
Make codebook for classcodes object
}
\examples{
# All codes from ICD-10-CM used by Elixhauser
codebook(elixhauser, "icd10cm")

# All codes from ICD-9-CM Disease part used by Elixhauser enhanced version
codebook(elixhauser, "icd9cmd",
  cc_args = list(regex = "regex_icd9cm_enhanced")
)
}
\seealso{
codebooks

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{print.codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
