% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{codedata}
\alias{as.codedata}
\alias{codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ...)

is.codedata(x)
}
\arguments{
\item{x}{data frame with columns "id", "date" and "code" or
object of class \code{\link{pardata}}.}

\item{...}{possible additional objects to merge with \code{x}}
}
\value{
\code{as.codedata} returns a data frame
  with mandatory columns:
\describe{
\item{id}{individual id}
\item{date}{date when code were valid}
\item{code}{code}
}
Additional columns might exist (preserved from \code{x})

\code{is.codedata} returns \code{TRUE} if these same conditions are met and
\code{FALSE} otherwise.
(Note that \code{codedata} is not a formal class of its own!)
}
\description{
Code data
}

