% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{codedata}
\alias{codedata}
\alias{as.codedata}
\alias{is.codedata}
\alias{as.codedata.data.table}
\alias{as.codedata.pardata}
\title{Code data}
\usage{
as.codedata(x, ...)

is.codedata(x)

\method{as.codedata}{data.table}(x, ..., setkeys = FALSE)

\method{as.codedata}{pardata}(x, ...)
}
\arguments{
\item{x}{data frame with columns "id", "date" and "code" or
object of class \code{\link{pardata}}.}

\item{...}{by default, codes for future dates or dates before "1970-01-01"
are ignored (with a warning). Specify \code{to, from}
(as passed to \code{\link{filter_dates}}) to override.}

\item{setkeys}{should keys be set for the data.table object?
To initially set the key for a large data set can be slow
but it could be beneficial for later use of the object.}
}
\value{
\code{as.codedata} returns a data frame
  with mandatory columns:
\describe{
\item{id}{individual id}
\item{date}{date when code were valid}
\item{code}{code}
}
Additional columns might exist (preserved from \code{x})

\code{is.codedata} returns \code{TRUE} if these same conditions are met and
\code{FALSE} otherwise.
(Note that \code{codedata} is not a formal class of its own!)
}
\description{
Code data
}
\examples{

x <- data.frame(id = 1, date = as.Date("2017-02-02"), code = "a")
as.codedata(x)

# Drop dates outside specified limits
y <- data.frame(id = 2, date = as.Date("3017-02-02"), code = "b")
z <- rbind(x, y)
as.codedata(z)

}
