% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedata.R
\name{codedata}
\alias{as.codedata}
\alias{codedata}
\alias{is.codedata}
\title{Code data}
\usage{
as.codedata(x, ...)

is.codedata(x)
}
\arguments{
\item{x}{object}

\item{...}{possible additional objects to merge with \code{x}}
}
\value{
\code{as.codedata} returns an object of class \code{tbl_df} with mandatory columns:
\describe{
\item{id}{individual id}
\item{date}{date when code were valid}
\item{code}{code}
}
Additional columns might exist (preserved from \code{x})
\code{is.codedata} reuturns \code{TRUE} if these same conditions are met,
\code{FALSE} otherwise.
(Note that \code{codedata} is not a formal class of its own!)
}
\description{
Code data
}
\examples{
\dontrun{
as.codedata(oppen, sluten)
}
}

