% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder-package.R
\docType{package}
\name{coder-package}
\alias{coder}
\alias{coder-package}
\title{coder: Deterministic Categorization of Items Based on External Code Data}
\description{
Fast categorization of items based on external code data identified by 
  regular expressions.
  It is easy to introduce new classification schemes (by 'classcodes' objects) or  
  to use default schemes included in the package. Use cases includes patient 
  categorization based on co-morbidity indices such as Charlson, Elixhauser, 
  RxRisk V, or the co-morbidity-polypharmacy score (CPS), as well as adverse 
  events after arthroplasty surgery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eribul/coder}
  \item Report bugs at \url{https://github.com/eribul/coder/issues}
}

}
\author{
\strong{Maintainer}: Erik Bulow \email{eriklgb@gmail.com} (\href{https://orcid.org/0000-0002-9973-456X}{ORCID})

}
\keyword{internal}
