% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder-package.R
\docType{package}
\name{coder-package}
\alias{coder}
\alias{coder-package}
\title{coder: Deterministic Categorization of Items Based on External Code Data}
\description{
Identify items and classify them into categories by coded data.
  Functionality of the package relies on three central objects.
  1) A data frame ('data.table') with item data (for example patients) to classify.
  2) A data frame ('data.table') with codes (for example 'ICD' 10-codes) for each item
  valid at distinct time points.
  3) A classification scheme (a 'classcodes' object) defining the
  classification based on regular expressions.
  It is easy to introduce new classification schemes to use with arbitrary
  code data and some default 'classcodes' objects are included in the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eribul/coder}
  \item Report bugs at \url{https://github.com/eribul/coder/issues}
}

}
\author{
\strong{Maintainer}: Erik Bulow \email{eriklgb@gmail.com} (\href{https://orcid.org/0000-0002-9973-456X}{ORCID})

}
\keyword{internal}
