% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify.R
\name{codify}
\alias{codify}
\title{Codify elements (within period)}
\usage{
codify(
  data,
  codedata,
  id,
  code,
  date = NULL,
  code_date = NULL,
  days = NULL,
  alnum = FALSE,
  .copy = NA
)
}
\arguments{
\item{data}{\link{data.frame} or \link[data.table:data.table]{data.table::data.table} with at least case id
(\code{character}), and optional date (\code{\link{Date}}) of interest}

\item{codedata}{\link{data.frame} or \link[data.table:data.table]{data.table::data.table} with columns
including case id (\code{character}), code and an optional date (\link{Date}) for
each code. An optional column \code{condition} might distinguish codes/dates
with certain characteristics (see example).}

\item{id, code, date, code_date}{column names with case id
(\code{character} from \code{data} and \code{codedata}), \code{code} (from \code{data}) and
optional date (\link{Date} from \code{data}) and
\code{code_date} (\link{Date} from \code{codedata}).}

\item{days}{numeric vector of length two with lower and upper bound for range
of relevant days relative to \code{date}. See "Relevant period".}

\item{alnum}{Should codes be cleaned from all non alphanumeric characters?}

\item{.copy}{Should the object be copied internally by \code{\link[data.table:copy]{data.table::copy()}}?
\code{NA} (by default) means that objects smaller than 1 GB are copied.
If the size is larger, the argument must be set explicitly. Set \code{TRUE}
to make copies regardless of object size. This is recommended if enough RAM
is available. If set to \code{FALSE}, calculations might be carried out
but the object will be changed by reference.}
}
\value{
Object of class \code{codified} (inheriting from \link[data.table:data.table]{data.table::data.table}).
Essentially \code{data} with additional columns:
\verb{code, code_date}: left joined from \code{codedata} or \code{NA}
if no match within period. \code{in_period}: Boolean indicator if the case
had at least one code within the specified period.

The output has one row for each combination of "id" from \code{data} and
"code" from \code{codedata}. Rows from \code{data} might be repeated
accordingly.
}
\description{
Enhance case data with code data, possibly limited to relevant period.
}
\section{Relevant period}{

Some examples for argument \code{days}:
\itemize{
\item \code{c(-365, -1)}: window of one year prior to the \code{date}
column of \code{data}. Useful for patient co-morbidity.
\item \code{c(1, 30)}: window of 30 days after \code{date}.
Useful for adverse events after a surgical procedure.
\item \code{c(-Inf, Inf)}: no limitation on non missing dates.
\item \code{NULL}: no time limitation at all.
}
}

\examples{
codify(ex_people, ex_icd10, id = "name", code = "icd10",
  date = "surgery", code_date = "admission", days = c(-365, 0))
}
\seealso{
Other verbs: 
\code{\link{categorize}()},
\code{\link{classify}()},
\code{\link{index_fun}}
}
\concept{verbs}
