% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify.R
\name{codify}
\alias{codify}
\title{codify elements}
\usage{
codify(x, from, id = "id", date, years = 1)
}
\arguments{
\item{x}{data.frame with at least two columns, one with case (patient)
identification (column name specified by argument \code{id}) and one with a
date of interest (column name specified by argument \code{date})}

\item{from}{object with code data for which \code{\link{is.codedata}} is \code{TRUE}}

\item{id}{name of column in \code{x} containing case (patient) identification}

\item{date}{name of column in \code{x} with date of interest}

\item{years}{number of years before \code{date} during which codes from
\code{from} are relevant}
}
\value{
Object of class \code{tbl_df} with columns corresponding to \code{x}
and additional columns matched from \code{from}:
\itemize{
  \item \code{code}: code as matchaed from \code{from} (\code{NA} if no match within period)
  \item \code{code_date}: corresponding date for which the code was valid
        (\code{NA} if no match within period)
  \item \code{in_period}: boolean indicator if the unit (patient) had at least one
  code within the specified period
}
 Ther output has at one row for each combination of unit and code.
}
\description{
codify elements
}
\examples{
codify(ex_people, ex_icd10, id = "name", date = "surgery")
}

