% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify.R
\name{codify}
\alias{codify}
\title{codify elements}
\usage{
codify(x, from, id = "id", date = NULL, days = NULL, .copy = NA)
}
\arguments{
\item{x}{data.frame with at least two columns, one with case (patient)
identification (column name specified by argument \code{id}) and one with a
date of interest (column name specified by argument \code{date})}

\item{from}{object with code data for which \code{\link{is.codedata}} is
\code{TRUE}}

\item{id}{name of column in \code{x} containing case (patient) identification}

\item{date}{name of column in \code{x} with possible date of interest
(\code{NULL} by default but must be specified if argument 'days' is
used).}

\item{days}{numeric vector of length two with lower and upper bound of range
of days relative to \code{date} for which codes from \code{from} are
relevant. (For example \code{c(-365, -1)} implies a time window of one year
prior to \code{date}, which might be useful for calculating comorbidity
indices, while \code{c(1, 30)} gives a window of 30 days after \code{date},
which might be used for calculating adverse events after a surgical
procedure.) \code{c(-Inf, Inf)} means no limitation on non missing dates.
\code{NULL} means no time limitation at all.}

\item{.copy}{Should the object be copied internally by \code{\link{copy}}?
\code{NA} (by default) means that objects smaller than 1 Gb are copied.
If the size is larger, the argument must be set explicitly. Set \code{TRUE}
to make copies regardless of object size. This is recomended if enough RAM
is available. If set to \code{FALSE}, calculations might be carried out
but the object will be changed by reference.}
}
\value{
Data frame (\code{data.table}) with columns corresponding to \code{x}
  and additional columns matched from \code{from}: \itemize{ \item
  \code{code}: code as matched from \code{from} (\code{NA} if no match
  within period) \item \code{code_date}: corresponding date for which the
  code was valid (\code{NA} if no match within period) \item
  \code{in_period}: boolean indicator if the unit (patient) had at least one
  code within the specified period } The output has at one row for each
  combination of "id" and code. Note that other columns of \code{x} might be
  repeated accordingly.
}
\description{
codify elements
}
\examples{
codify(ex_people, ex_icd10, id = "name", date = "surgery", days = c(-365, 0))
}
\seealso{
Other verbs: \code{\link{add}}, \code{\link{classify}},
  \code{\link{index_fun}}
}
\concept{verbs}
