% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{dates_within}
\alias{dates_within}
\title{Check if dates are within limits}
\usage{
dates_within(x, from = "1970-01-01", to = Sys.Date())
}
\arguments{
\item{x}{Date vector}

\item{from}{lower date limit to filter out irrelevant dates
(\code{1970-01-01} by default, since we assume that little data was recorded
before 'epoch'; this is a bold assumption, however, which should be
explicitly re-considered for relevant applications)}

\item{to}{upper date limit to filter out irrelevant dates
(the current data from \code{\link[=Sys.Date]{Sys.Date()}}, by default,
since we assume that future events are unknown,
thus indicating coding errors)}
}
\value{
Logical vector with \code{TRUE} if date within limits, \code{FALSE} otherwise.
}
\description{
Check if dates are within limits
}
\examples{

#                 valid         future      too early
x <- as.Date(c("2017-02-02", "2050-02-02", "1969-02-02"))
coder:::dates_within(x) # TRUE FALSE FALSE
coder:::dates_within(x, from = "2000-01-01", to = "2100-01-01") #  TRUE  TRUE FALSE
}
\seealso{
Other helper: 
\code{\link{as.keyvalue.classcodes}()},
\code{\link{filter_dates}()},
\code{\link{spss2date}()}
}
\concept{helper}
\keyword{internal}
