% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{elix_icd10}
\alias{elix_icd10}
\title{Classcodes for Elixhauser based on ICD-10 codes}
\format{A data frame with 31 rows and 2 variables:
\describe{
  \item{group}{comorbidity groups}
  \item{regex}{regular expressions identifying ICD-10 codes of each group}
  \item{sum_all}{all weights = 1. This is included for convenience when
  calculating several indices simultainisly}
  \item{walraven}{weights suggested by Walraven et al (2009)}
  \item{sid29}{weights suggested by Thompson et al. (2015)
    based on all conditions except cardian arryrtmias}
  \item{sid30}{weights suggested by Thompson et al. (2015)
    based on all conditions}
}}
\source{
Quan Hude et al. (2005). Coding algorithms for defining
  comorbidities in ICD-9-CM and ICD-10 administrative data.
  Medical care, 1130-1139.
  \url{www.jstor.org/stable/3768193}

Walraven, C. Van, Austin, P. C., Jennings, A., Quan, H., Alan, J., Walraven,
  C. Van, … Jennings, A. (2009).
  A Modification of the Elixhauser Comorbidity Measures Into a Point System
  for Hospital Death Using Administrative Data. Medical Care, 47(6), 626–633.

Thompson, N. R., Fan, Y., Dalton, J. E., Jehi, L., Rosenbaum, B. P.,
  Vadera, S., & Griffith, S. D. (2015).
  A new Elixhauser-based comorbidity summary measure to predict in-hospital
  mortality. Med Care, 53(4), 374–379.
  http://doi.org/10.1097/MLR.0000000000000326
}
\usage{
elix_icd10
}
\description{
A \code{\link{classcodes}} object to use with \code{\link{classify}}
}
\seealso{
Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}},
  \code{\link{hip_adverse_events_icd10_old}},
  \code{\link{hip_fracture_ae_icd10}},
  \code{\link{hip_fracture_ae_kva}},
  \code{\link{knee_adverse_events_icd10}},
  \code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
