% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{ex_icd10}
\alias{ex_icd10}
\title{Example data for random codes assigned to random people}
\format{Data frames with 1,000 rows and 4 variables:
\describe{
  \item{id}{Random names corresponding to column \code{name} in dataset
    \code{ex_people}}
  \item{date}{random dates corresponding to registered (comorbidity) codes}
  \item{code}{(comorbidity) codes as given by ICD-10}
  \item{hdia}{boolean marker if corresponding code is the main diagnose of
    the hospital visit (randomly assigned to 10 percent of the codes)}
}}
\usage{
ex_icd10
}
\description{
Example data for fictive ICD-10-diagnoses to use for testing and
in examples.
}
\seealso{
ex_people
}
\keyword{datasets}

