% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{ex_people}
\alias{ex_people}
\title{Example data for random people}
\format{
Data frames with 100 rows and 2 variables:
\describe{
  \item{name}{random person names}
  \item{surgery}{random dates for performed surgery on each patient}
}
}
\usage{
ex_people
}
\description{
Example data for fictive people to use for testing and in examples.
}
\seealso{
Other example data: 
\code{\link{ex_carbrands}},
\code{\link{ex_cars}},
\code{\link{ex_icd10}}
}
\concept{example data}
\keyword{datasets}
