% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dates.R
\name{filter_dates}
\alias{filter_dates}
\alias{filter_dates.data.frame}
\alias{filter_dates.Date}
\title{Filter dates outside specified limits}
\usage{
filter_dates(x, ...)

\method{filter_dates}{data.frame}(x, ...)

\method{filter_dates}{Date}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{object to filter}

\item{...}{arguments passed to \code{\link{dates_within}}.
It is here useful to specify the \code{limits} argument}

\item{na.rm}{remove dates outside limits completely (otherwise keep position
in vector but set to \code{NA})}
}
\value{
Object of same class as \code{x}. If date vector, a filtered vector,
if data frame (with one column named "date"), the same data frame but only
with rows with dates within limits.
}
\description{
Filter dates outside specified limits
}
\examples{
#                 valid         future      too early
x <- as.Date(c("2017-02-02", "2050-02-02", "1969-02-02"))
filter_dates(x) # "2017-02-02" NA           NA
filter_dates(x, na.rm = TRUE) # "2017-02-02"

filter_dates(
  data.frame(
    date = x,
    foo  = 1:3
  )
)
}
\seealso{
Other helper: 
\code{\link{dates_within}()},
\code{\link{spss2date}()}
}
\concept{helper}
