% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_classcodes}
\alias{get_classcodes}
\title{Get classcodes object}
\usage{
get_classcodes(x, from = NULL)
}
\arguments{
\item{x}{classcodes specification as either a name or a classcodes object,
or a classcodes object itself.}

\item{from}{object that classcodes could be inherited from}
}
\value{
\code{\link{classcodes}} object or \code{NULL} if no object found.
}
\description{
Get classcodes object
}
\seealso{
Other classcodes: \code{\link{all_classcodes}},
  \code{\link{as.data.frame.classified}},
  \code{\link{classcodes}},
  \code{\link{print.summary.classcodes}},
  \code{\link{summary.classcodes}}, \code{\link{visualize}}
}
\concept{classcodes}
\keyword{internal}
