% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{hip_ae_hailer}
\alias{hip_ae_hailer}
\title{Classcodes for infection and dislocation after hip arthroplasty}
\format{
Data frame with 3 columns:
\describe{
\item{group}{Infection or dislocation}
\item{icd10}{regular expressions based on ICD-10}
\item{kva}{regular expressions based on NOMESCO/KVA codes}
}
}
\usage{
hip_ae_hailer
}
\description{
Classcodes for infection and dislocation after hip arthroplasty
}
\seealso{
ae

Other default classcodes: 
\code{\link{ae}},
\code{\link{charlson}},
\code{\link{cps}},
\code{\link{elixhauser}},
\code{\link{rxriskv}}
}
\concept{default classcodes}
\keyword{datasets}
