% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Calculate index based on classification scheme}
\usage{
index(x, classcodes = NULL)
}
\arguments{
\item{x}{matrix as output from \code{classify}}

\item{classcodes}{object of class \code{classcodes}.
Needed if classes have different weights. \code{NULL} (as by default)
gives index = 1 if element belongs to any class, 0 otherwise.}
}
\value{
Named numeric index vector with names corresponding to \code{rownames(x)}
}
\description{
Calculate index based on classification scheme
}
\examples{
x <- classify(c("C80", "I20", "XXX"))
index(x)
}

