% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_fun.R
\name{index_fun}
\alias{index_fun}
\alias{index}
\alias{index.data.frame}
\alias{index.matrix}
\title{Calculate index based on classification scheme}
\usage{
index(classified, ...)

\method{index}{data.frame}(classified, ...)

\method{index}{matrix}(classified, index = NULL, cc = NULL, ...)
}
\arguments{
\item{classified}{output from \code{\link[=classify]{classify()}}}

\item{...}{used internally}

\item{index}{name of column with 'weights' from corresponding
\code{\link{classcodes}} object. Can be \code{NULL} if the index is just a
count of relevant classes.}

\item{cc}{\code{\link{classcodes}} object. Can be \code{NULL} if information
already present as attribute of \code{classified} (which is often the case)
and/or if index calculated without weights.}
}
\value{
Named numeric index vector with names corresponding to
\code{rownames(x)}
}
\description{
(Weighted) sum of all identified classes for each case.
Index weights for subordinate hierarchical classes
(as identified by \code{attr(cc, "hierarchy")}) are excluded in presence of
superior classes if index specified with argument \code{index}.
}
\examples{

# Prepare some codified data with ICD-10 codes during 1 year (365 days)
# before surgery
x <-
  codify(
    ex_people,
    ex_icd10,
    id        = "name",
    code      = "icd10",
    date      = "surgery",
    days      = c(-365, 0),
    code_date = "admission"
  )

# Classify those patients by the Charlson and Elixhasuer comorbidity indices
cl <- classify(x, "charlson")

# Calculate (weighted) index values
index(cl)                  # Un-weighted sum/no of conditions for each patient
index(cl, "quan_original") # Weighted index (Quan et al. 2005; see `?charlson`)
index(cl, "quan_updated")  # Weighted index (Quan et al. 2011; see `?charlson`)

# Tabulate index for all patients.
# As expected, most patients are healthy and have index = 0/NA,
# where NA indicates no recorded hospital visits
# found in `ex_icd10` during codification.
# In practice, those patients might be assumed to have 0 comorbidity as well.
table(index(cl, "quan_original"), useNA = "always")

# If `cl` is a matrix without additional attributes (as imposed by `codify()`)
# an explicit classcodes object must be specified by the `cc` argument
cl2 <- as.matrix(cl)
index(cl2, cc = "charlson")
}
\seealso{
Other verbs: 
\code{\link{categorize}()},
\code{\link{classify}()},
\code{\link{codify}()}
}
\concept{verbs}
