% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_fun.R
\name{index_fun}
\alias{index_fun}
\alias{index}
\alias{index.data.frame}
\alias{index.matrix}
\title{Calculate index based on classification scheme}
\usage{
index(x, ...)

\method{index}{data.frame}(x, ...)

\method{index}{matrix}(x, by = NULL, from = NULL, ...)
}
\arguments{
\item{x}{output from \code{classify}
(or possibly from \code{\link{as.data.frame.classified}})}

\item{...}{used internally}

\item{by}{name of column with 'weights' from corresponding
\code{\link{classcodes}} object. Can be omitted if the index is just the
count of relevant classes.}

\item{from}{\code{\link{classcodes}} object. Can be omitted if information
already present as attribute of \code{x} (which is often the case) and/or
if index calculated without weights.}
}
\value{
Named numeric index vector with names corresponding to
  \code{rownames(x)}
}
\description{
Calculate index based on classification scheme
}
\examples{

# Calculate Elixhauser comorbidity index
cl <- classify(c("C80", "I20"), "elix_icd10")
index(cl)

# Calculate Charlson-index using original weights
cl <- classify(c("C80", "I21"), "charlson_icd10")
index(cl, "quan_original")

# Calculate Charlson-index using updated weights
cl <- classify(c("C80", "I21"), "charlson_icd10")
index(cl, "quan_updated")


# Find patients with adverse events after hip surgery
co <- codify(ex_people, ex_icd10, id = "name",
    date = "surgery", days = c(-365, 0))
cl <- classify(co, "hip_adverse_events_icd10_old")
index(cl)

}
\seealso{
Other verbs: \code{\link{add}}, \code{\link{classify}},
  \code{\link{codify}}
}
\concept{verbs}
