% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_fun.R
\name{index_fun}
\alias{index_fun}
\alias{index}
\alias{index.data.frame}
\alias{index.matrix}
\title{Calculate index based on classification scheme}
\usage{
index(classified, ...)

\method{index}{data.frame}(classified, ...)

\method{index}{matrix}(classified, index = NULL, cc = NULL, ...)
}
\arguments{
\item{classified}{output from \code{classify}}

\item{...}{used internally}

\item{index}{name of column with 'weights' from corresponding
\code{\link{classcodes}} object. Can be \code{NULL} if the index is just a
count of relevant classes.}

\item{cc}{\code{\link{classcodes}} object. Can be \code{NULL} if information
already present as attribute of \code{classified} (which is often the case) and/or
if index calculated without weights.}
}
\value{
Named numeric index vector with names corresponding to
  \code{rownames(x)}
}
\description{
Calculate index based on classification scheme
}
\examples{

# Calculate Elixhauser comorbidity index
cl <- classify(c("C80", "I20"), "elixhauser")
index(cl)

# Calculate Charlson-index using original weights
cl <- classify(c("C80", "I21"), "charlson")
index(cl, "quan_original")

# Calculate Charlson-index using updated weights
cl <- classify(c("C80", "I21"), "charlson")
index(cl, "quan_updated")

}
\seealso{
Other verbs: 
\code{\link{categorize}()},
\code{\link{classify}()},
\code{\link{codify}()}
}
\concept{verbs}
