% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R, R/par.R
\docType{data}
\name{pardata}
\alias{as.pardata}
\alias{is.pardata}
\alias{pardata}
\title{SHPR matched PAR-data 2001-2013}
\format{A data frame with 9,844,928 rows and 3 variables:
\describe{
  \item{lpnr}{anonymised patient identification number}
  \item{icd_date}{date for registration of ICD10 code}
  \item{icd}{ICD10-code}
}}
\usage{
pardata

as.pardata(x)

is.pardata(x)
}
\arguments{
\item{x}{data.frame with par-data}
}
\value{
\code{as.pardata} returns an object of class "pardata".
\code{is.pardata} returns \code{TRUE} if object is of class "pardata",
\code{FALSE} otherwise.
}
\description{
A dataset containing ICD10-codes for SHPR matched patients from
the Patient Administratice Register.

Functions to check that a data set contains patient data from
the Swedish Patient Register (PAR).
}
\keyword{datasets}

