% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{pardata}
\alias{as.pardata}
\alias{is.pardata}
\alias{pardata}
\title{Data from the Swedish Patient Register (PAR).}
\usage{
as.pardata(x)

is.pardata(x)
}
\arguments{
\item{x}{data.frame with par-data}
}
\value{
\code{as.pardata} returns an object of class "pardata".
\code{is.pardata} returns \code{TRUE} if object is of class "pardata",
\code{FALSE} otherwise.
}
\description{
Functions to check that a data set contains patient data from
the Swedish Patient Register (PAR).
}

