% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{print.summary.classcodes}
\alias{print.summary.classcodes}
\title{Print summary for classcodes object}
\usage{
\method{print}{summary.classcodes}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{summary.classcodes}}}

\item{...}{ignored}
}
\value{
Nothing. This function is called for its side effects
}
\description{
Print summary for classcodes object
}
\examples{
print(summary(ex_carbrands))
}
