% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{print.summary.classcodes}
\alias{print.summary.classcodes}
\title{Print summary for classcodes object}
\usage{
\method{print}{summary.classcodes}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[=summary.classcodes]{summary.classcodes()}}}

\item{...}{arguments to control the printing of a tibble}
}
\value{
Nothing. This function is called for its side effects
}
\description{
Print summary for classcodes object
}
\examples{
print(summary(elixhauser, coding = "icd10cm"))
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.codebook}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
