% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{print.summary.classcodes}
\alias{print.summary.classcodes}
\title{Print summary for classcodes object}
\usage{
\method{print}{summary.classcodes}(x, n_print = 50, ...)
}
\arguments{
\item{x}{object of class \code{\link{summary.classcodes}}}

\item{n_print}{maximum characters to print from the \code{codes} column}

\item{...}{ignored}
}
\value{
Nothing. This function is called for its side effects
}
\description{
Print summary for classcodes object
}
\examples{
print(summary(elixhauser, coding = "icd10cm"))
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
