% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{print.summary.classcodes}
\alias{print.summary.classcodes}
\title{Print summary for classcodes object}
\usage{
\method{print}{summary.classcodes}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[=summary.classcodes]{summary.classcodes()}}}

\item{...}{
  Arguments passed on to \code{\link[tibble:NA]{tibble:::print.tbl}}
  \describe{
    \item{\code{n}}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than option \code{tibble.print_max}. Otherwise, will print
\code{tibble.print_min} rows.}
    \item{\code{width}}{Width of text output to generate. This defaults to \code{NULL}, which
means use \code{getOption("tibble.width")} or (if also \code{NULL})
\code{getOption("width")}; the latter displays only the columns that fit on one
screen. You can also set \code{options(tibble.width = Inf)} to override this
default and always print all columns.}
    \item{\code{n_extra}}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the default,
will print information about at most \code{tibble.max_extra_cols} extra columns.}
  }}
}
\value{
Nothing. This function is called for its side effects
}
\description{
Print summary for classcodes object
}
\examples{
print(summary(elixhauser, coding = "icd10cm"))
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
