% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{rxriskv}
\alias{rxriskv}
\alias{rxriskv_atc}
\alias{rxriskv.modified_atc}
\title{Classcodes for RxRiskV (original and modified)}
\format{Data frames with 39 rows and 2 variables:
\describe{
  \item{group}{disease group}
  \item{regex}{regular expressions identifying ATC codes of each group}
}}
\usage{
rxriskv_atc

rxriskv.modified_atc
}
\description{
\code{\link{classcodes}} object to use with \code{\link{classify}}
}
\details{
\code{rxriskv.modified_atc} adopted by Anne Garland to better mirror
 medical use in Sweden.
}
\seealso{
Other default classcodes: \code{\link{ae}},
  \code{\link{charlson_icd10}}, \code{\link{cps_icd10}},
  \code{\link{elix_icd10}}
}
\concept{default classcodes}
\keyword{datasets}
