% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_for_datasets.R
\docType{data}
\name{rxriskv}
\alias{rxriskv}
\alias{rxriskv_icd10}
\alias{rxriskv_modified_icd10}
\title{Classcodes for RxRiskV (original and modified)}
\format{Data frames with 39 rows and 2 variables:
\describe{
  \item{group}{disease group}
  \item{regex}{regular expressions identifying ATC codes of each group}
}}
\usage{
rxriskv_icd10

rxriskv_modified_icd10
}
\description{
\code{\link{classcodes}} object to use with \code{\link{classify}}
}
\seealso{
Other default classcodes: \code{\link{charlson_icd10}},
  \code{\link{cps_icd10}}, \code{\link{elix_icd10}},
  \code{\link{hip_adverse_events_icd10_old}},
  \code{\link{knee_adverse_events_icd10}},
  \code{\link{tha_fracture_ae_icd10}},
  \code{\link{tha_fracture_ae_kva}}
}
\keyword{datasets}
