% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_classcodes.R
\name{set_classcodes}
\alias{set_classcodes}
\title{Set classcodes object}
\usage{
set_classcodes(
  cc,
  classified = NULL,
  regex = NULL,
  start = TRUE,
  stop = FALSE,
  tech_names = FALSE
)
}
\arguments{
\item{cc}{\code{\link{classcodes}} object or name of such}

\item{classified}{object that classcodes could be inherited from}

\item{regex}{name of column with regular expressions to use for
classification, either with or without prefix \code{regex_}.
\code{NULL} (default) uses the first classcodes column with prefix "regex".
This should be a sensible choice for default classcodes object included
in the package.}

\item{start, stop}{should codes start/end with the specified regular
expressions? If \code{TRUE}, column "regex" is prefixed/suffixed
by \verb{^/$}.}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{by}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.}
}
\value{
\code{\link{classcodes}} object.
}
\description{
Set classcodes object
}
\examples{
# Prepare a classcodes object for the Charlson comorbidity classification
# based on the default regular expressions
set_classcodes(charlson)

# Same as above but based on regular expressions for ICD-8
set_classcodes(charlson, regex = "regex_icd8_brusselaers")
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
