% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_classcodes.R
\name{set_classcodes}
\alias{set_classcodes}
\title{Set classcodes object}
\usage{
set_classcodes(
  cc,
  classified = NULL,
  regex = NULL,
  start = TRUE,
  stop = FALSE,
  tech_names = NULL
)
}
\arguments{
\item{cc}{\code{\link{classcodes}} object (or name of a default object from
\code{\link[=all_classcodes]{all_classcodes()}}).}

\item{classified}{object that classcodes could be inherited from}

\item{regex}{name of column with regular expressions to use for
classification.
\code{NULL} (default) uses \code{attr(obj, "regexpr")[1]}.}

\item{start, stop}{should codes start/end with the specified regular
expressions? If \code{TRUE}, column "regex" is prefixed/suffixed
by \verb{^/$}.}

\item{tech_names}{should technical column names be used? If \code{FALSE},
colnames are taken directly from group names of \code{cc}, if \code{TRUE},
these are changed to more technical names avoiding special characters and
are prefixed by the name of the classification scheme.
\code{NULL} (by default) preserves previous names if \code{cc} is inherited from
\code{classified} (fall backs to \code{FALSE} if not already set).}
}
\value{
\code{\link{classcodes}} object.
}
\description{
Prepare a \code{classcodes}object by specifying the regular expressions
to use for classification.
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebook}()},
\code{\link{print.classcodes}()},
\code{\link{print.classified}()},
\code{\link{summary.classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
