% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{summary.classcodes}
\alias{summary.classcodes}
\title{Summarizing a classcodes object}
\usage{
\method{summary}{classcodes}(object, regex = "regex", coding, ...)
}
\arguments{
\item{object}{classcodes object}

\item{regex}{name of column with regular expressions to use for
classification, either with or without prefix \code{regex_}}

\item{coding}{name (as character) of classification used for coding
(for example "icd10" or "atc")}

\item{...}{ignored}
}
\value{
List (invisable) with objects:

\itemize{
  \item{\code{object}: input \code{object}}
  \item{\code{summary}: a data frame with columns:
     \itemize{
    \item{\code{group}: Groups identified by \code{object}.}
    \item{\code{n}: The number oif codes to be recognized for each group.}
    \item{\code{codes}: Individual codes within each group.}
    }
  }
  \item{\code{coding}: input \code{coding}}
}
}
\description{
Classification schemes are formalized by regular expressions within the
classcodes objects. These are computationally effective but sometimes hard to
grasp. Use this function to present all individual codes identified for each
group in the classification.
}
\examples{
summary(elixhauser, coding = "icd10")
}
\seealso{
\code{\link{visualize}} for a graphical representation of the
  classcodes objects.

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{get_classcodes}()},
\code{\link{print.summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
