% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{summary.classcodes}
\alias{summary.classcodes}
\title{Summarizing a classcodes object}
\usage{
\method{summary}{classcodes}(object, ...)
}
\arguments{
\item{object}{classcodes object}

\item{...}{ignored}
}
\value{
Data frame (nivisable) with columns:
\describe{
  \item{group}{Groups identified by \code{object}.}
  \item{n}{The number oif codes to be recognized for each group.}
  \item{codes}{Individual codes within each group.}
}
}
\description{
Classification schemes are formalized by regular expressions within the
classcodes objects. These are computationally effective but sometimes hard to
grasp. Use this function to present all individual codes identified for each
group in the classification.
}
\examples{
summary(ex.carbrands_excars)
}
\seealso{
\code{\link{visualize}} for a graphical representation of the
  classcodes objects.

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{get_classcodes}()},
\code{\link{print.summary.classcodes}()},
\code{\link{visualize}()}
}
\concept{classcodes}
