% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.classcodes.R
\name{summary.classcodes}
\alias{summary.classcodes}
\title{Summarizing a classcodes object}
\usage{
\method{summary}{classcodes}(object, coding, ..., cc_args = list())
}
\arguments{
\item{object}{classcodes object}

\item{coding}{either a vector with codes from the original classification,
or a name (character vector of length one) keyvalue object from package
"decoder" (for example "icd10cm" or "atc")}

\item{...}{ignored}

\item{cc_args}{List of named arguments passed to \code{\link[=set_classcodes]{set_classcodes()}}}
}
\value{
List (invisible) with objects:
\itemize{
\item \code{object}: input \code{object}
\item \verb{summary:} a data frame with columns:
\itemize{
\item \code{group}: Groups identified by \code{object}.
\item \code{n}: The number of codes to be recognized for each group.
\item \code{codes}: Individual codes within each group.
}
\item \code{coding}: input \code{coding}
}
}
\description{
Classification schemes are formalized by regular expressions within the
classcodes objects. These are computationally effective but sometimes hard to
grasp. Use this function to present all individual codes identified for each
group in the classification.
}
\examples{
summary(elixhauser, coding = "icd10cm")
}
\seealso{
\code{\link[=visualize]{visualize()}} for a graphical representation of the
classcodes objects.

Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{codebooks}()},
\code{\link{codebook}()},
\code{\link{print.classcodes}()},
\code{\link{print.codebook}()},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{visualize.classcodes}()}
}
\concept{classcodes}
