% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualise classification scheme in web browser}
\usage{
visualize(cc, group = NULL, show = TRUE, ...)
}
\arguments{
\item{cc}{\code{\link{classcodes}} object or name of such}

\item{group}{names (as character vector) of groups to visualise
(all groups if \code{NULL})}

\item{show}{should a visualisation be shown in a web browser.
Set to \code{FALSE} to just retrieve a URL for later use.}

\item{...}{
  Arguments passed on to \code{\link[=set_classcodes]{set_classcodes}}
  \describe{
    \item{\code{regex}}{name of column with regular expressions to use for
classification, either with or without prefix \code{regex_}}
  }}
}
\value{
URL to website with visualisation (invisible)
}
\description{
Classes are visualised by their regular expressions in a web browser.
The visualisation does not give any details on group names, conditions or
weights but might be useful both for understanding of a clasification scheme
in use, and during the creation and debugging of such.
}
\examples{
\dontrun{

 # How is depression classified according to Elixhauser?
 visualize("elixhauser", "depression")

 # Compare the two diabetes groups according to Charlson
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"))

 # Is this different from the "Royal College of Surgeons classification?
 # Yes, there is only one group for diabetes
 visualize("charlson",
   c("diabetes without complication", "diabetes complication"),
   regex = "rcs"
 )

 # Show all groups from Charlson
 visualize("charlson")

 # Get URL for later visualisation
 (visualize("hip_fracture_ae", show = FALSE))

}
}
\seealso{
Other classcodes: 
\code{\link{all_classcodes}()},
\code{\link{as.data.frame.classified}()},
\code{\link{classcodes}},
\code{\link{print.summary.classcodes}()},
\code{\link{set_classcodes}()},
\code{\link{summary.classcodes}()}
}
\concept{classcodes}
